Rename kernels/ to device/
