+++ mlx/backend/cuda/reduce/reduce_ops.cuh
@@ -2,7 +2,7 @@
 
 #pragma once
 
-#include "mlx/backend/cuda/kernels/utils.cuh"
+#include "mlx/backend/cuda/device/utils.cuh"
 
 namespace mlx::core::cu {
